+++ libyafuse2/libfat/super.c
@@ -95,7 +95,7 @@ int32_t fat_fill_sb(struct fat_super_block *sb)
   sz = sizeof(struct fat_boot_sector);
   ret = io_read((uint8_t *)&sb->bs, sz);
   if (ret != 0) {
-    memset((void *)&sb->bs, 0, sz);
+    memset((void *)&sb->bs, 0, (size_t)sz);
     return -1;
   }
 
@@ -129,7 +129,7 @@ int32_t fat_fill_sb(struct fat_super_block *sb)
   sz = sizeof(struct fat_boot_bsx);
   ret = io_read((uint8_t *)&sb->bb, sz);
   if (ret != 0) {
-    memset((void *)&sb->bb, 0, sz);
+    memset((void *)&sb->bb, 0, (size_t)sz);
     return -1;
   }
 
@@ -147,7 +147,7 @@ int32_t fat_fill_sb(struct fat_super_block *sb)
     sz = sizeof(struct fat_boot_fsinfo);
     ret = io_read((uint8_t *)&sb->bf, sz);
     if (ret != 0) {
-      memset((void *)&sb->bf, 0, sz);
+      memset((void *)&sb->bf, 0,(size_t)sz);
       return -1;
     }
